refactor: extract into exportService.ts
